requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Pydantic
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Async file operations
aiofiles>=23.2.1

# Image processing
Pillow>=10.2.0

# PDF processing
pdf2image>=1.16.0
pypdf>=3.17.0

# File type detection
python-magic>=0.4.27

# HTTP client (for future webhooks/callbacks)
httpx>=0.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
python-dotenv>=1.0.0
